Started homepage and header
