+++ src/components/common/Footer/footer.scss
@@ -1,0 +1,8 @@
+@import "style/common/constants";
+@import "style/common/typography";
+
+.footer {
+  background-color: $black-russian;
+  width: 100vw;
+  height: 12vh;
+}
