+++ src/style/common/_constants.scss
@@ -2,3 +2,4 @@
 $white: #ffff;
 $black-russian: #00031a;
 $red-thunderbird: #d42f1a;
+$catskill-white: #f4f7fa;
